requirements.txt
# Core scraping dependencies
playwright>=1.40.0
psycopg2-binary>=2.9.9
redis>=5.0.0

# FastAPI and API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Celery
celery>=5.3.0
flower>=2.0.0

# Elasticsearch
elasticsearch>=8.11.0

# Prometheus metrics
prometheus-client>=0.19.0

# Data processing (minimal)
pandas>=2.0.0
numpy>=1.24.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
